Task 16 second correction

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,15 +9,15 @@
  */
 int main(void)
 {
-	uintmax_t a, b, c, i;
+	unsigned long long a, b, c, i;
 
 	a = 1;
 	b = 2;
-	printf("%ju, %ju", a, b);
+	printf("%llu, %llu", a, b);
 	for (i = 2; i < 98; i++)
 	{
 		c = a + b;
-		printf(", %ju", c);
+		printf(", %llu", c);
 		a = b;
 		b = c;
 	}
